requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
apipkg==1.5               # via execnet
attrs==19.3.0             # via pytest
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
crc32c==2.0.1             # via -r requirements.in
execnet==1.7.1            # via pytest-xdist
idna==2.10                # via requests
kafka-python==2.0.1       # via -r requirements.in
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-forked==1.2.0      # via pytest-xdist
pytest-xdist==1.33.0      # via -r requirements.in
pytest==5.4.3             # via -r requirements.in, pytest-forked, pytest-xdist
requests==2.24.0          # via -r requirements.in
six==1.15.0               # via packaging, pytest-xdist
urllib3==1.25.10          # via requests
wcwidth==0.2.5            # via pytest
